refactor(eval): extract benchmark/comparison reporting into eval-benchmark.js
ARCH-3: moved the 14-function token-accounting + benchmark-generation + result-comparison
cluster (generateBenchmark, compareResults, generateRawBenchmarkData, metricsFromResults, ...)
out of eval.js into scripts/lib/eval-benchmark.js. One-directional dependency
(eval-benchmark -> eval), no cycle. cli.js + eval.test.js import the moved functions directly.
Source moved verbatim (no logic drift). eval.js 1481 -> 1077; eval-benchmark.js 447.

EVAL-6 (eval-graders.js split) intentionally NOT done: grader types cross-cut (gradeWithMixed
needs both model + behavioral paths) and ~400 lines of shared infra span every type — a by-type
split would need a 'common' dump module. Left intact per the low-priority don't-force-it guidance.
